Add timeout for accepting the same event again
On things that change rarely but we want to query it often, we have an
optimization that instructs a queue to only accept different events.
Sometimes we want to enqueue and upload the same event, just to make
sure that the device is still "alive".  This change add a timeout for
that.