Disable debug output even in debug builds
While the debug output was turned off in release build, SDK users who
are testing the SDK might run the SDK in debug mode and our debug output
is cluttering the log output.  So we should turn off the debug output in
debug build as well, and it should be turned on only if asked for.

To turn on the SDK debug output, you should define `SF_DEBUG_ENABLE`
macro in your build settings.