Let users choose the default queue identifier
Currently we provide a default queue to SDK users, but they might want
to tweak the configurations of the default queue.  However, it also
might be a good idea to keep the queue configurations immutable once the
queue is created.

So we could probably solve it this way: Let users create a new queue
with the configurations of their choice, and tell the SDK to treat that
queue as the default one (by setting the `defaultQueueIdentifier`
property of the `Sift` object).