Prepare for device properties and app state split
We will split data we collect into two groups: device properties and app
state.  This change is refactoring the relevant classes in preparation
for this upcoming split.

This change includes:

* Add `SFHtDictionary` class, which replaces `SFIosDeviceProperties`
  class (and will be used for app state, too).

* Define device property entry types in `SFIosDeviceProperties.m` (we
  will also have a similar `SFAppState.m`).

* Move device property collection code to `SFIosDeviceProperties.m`, and
  only when-to-collect logic remains in `SFDevicePropertiesReporter.m`.

* Remove device properties that we decide should belong to app state.

The `SFHtDictionary` is a heterogeneous-typed dictionary for storing
device properties (and will also be storing app state).

These device properties are removed because they should belong to app
state:
* network addresses
* battery state and level
* device orientation
* proximity state