Add app state schema
We will capture two kinds of snapshot: app state and device properties.
The app state snapshots are for frequent-changing data and the device
properties snapshots are for the rest.

This change includes schema for:
* Battery state and level.
* Device orientation.
* Proximity sensor state.
* Location and heading.
* Network addresses.
* Motion data.

The motion data are actually not just a snapshot (i.e., current reading
of the sensors) but a series of readings in the past since one single
reading is probably useless (unless we take very frequent snapshots, but
that has other problems).  So `SFCollectIosAppState` function will not
read motion sensor data; instead, motion data is left for other parts of
the SDK.