Add app state collector
When app state collector collects, it capture an app state snapshot.  At
the moment, these sources may trigger a collection:
* App enters foreground.
* App enters background.
* Window root view controller's title is changed.
(We might add more sources in the future.)

We have rate limits on collecting frequency.  Also, if there is no
collection for a period, we request a collection (you could think of
this as a lower bound of collection frequency).